тип данных был изменен на корректный

+++ profcomff_definitions/DM/rating.py
@@ -17,14 +17,14 @@ class DmLecturerCommentAct(Base):
     lecturer_first_name: Mapped[str | None] = mapped_column(comment="Имя преподавателя")
     lecturer_last_name: Mapped[str | None] = mapped_column(comment="Фамилия преподавателя")
     lecturer_middle_name: Mapped[str | None] = mapped_column(comment="Отчество преподавателя")
-    mark_weighted: Mapped[int] = mapped_column(comment="Взвешенная оценка преподавателя", default=0, server_default="0")
-    mark_kindness_weighted: Mapped[int] = mapped_column(
+    mark_weighted: Mapped[float] = mapped_column(comment="Взвешенная оценка преподавателя", default=0, server_default="0")
+    mark_kindness_weighted: Mapped[float] = mapped_column(
         comment="Взвешенная доброта преподавателя", default=0, server_default="0"
     )
-    mark_clarity_weighted: Mapped[int] = mapped_column(
+    mark_clarity_weighted: Mapped[float] = mapped_column(
         comment="Взвешенная понятность преподавателя", default=0, server_default="0"
     )
-    mark_freebie_weighted: Mapped[int] = mapped_column(
+    mark_freebie_weighted: Mapped[float] = mapped_column(
         comment="Взвешенная халявность преподавателя", default=0, server_default="0"
     )
     rank: Mapped[int] = mapped_column(comment="Место в рейтинге", default=0, server_default="0")
